koko-ssh/requirements.txt
autopep8==1.6.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
cryptography==36.0.1
idna==3.3
importlib-metadata==4.10.1
paramiko==2.9.2
prettytable==3.0.0
pycodestyle==2.8.0
pycparser==2.21
PyNaCl==1.5.0
PyYAML==6.0
requests==2.27.1
six==1.16.0
toml==0.10.2
typing-extensions==4.0.1
urllib3==1.26.8
wcwidth==0.2.5
zipp==3.7.0
